requirements.txt
fastapi
uvicorn
biopython
psycopg2-binary
sqlalchemy
python-multipart
requests
reportlab